+++ dfir_rs/src/compiled/push/filter.rs
@@ -1,0 +1,50 @@
+use std::pin::Pin;
+use std::task::{Context, Poll};
+
+use futures::sink::Sink;
+use pin_project_lite::pin_project;
+
+pin_project! {
+    /// Same as [`std::iterator::Filter`] but as a [`Sink`].
+    ///
+    /// Synchronously filters items and sends the outputs to the following sink.
+    #[must_use = "sinks do nothing unless polled"]
+    pub struct Filter<Si, Func> {
+        #[pin]
+        sink: Si,
+        func: Func,
+    }
+}
+
+impl<Si, Func> Filter<Si, Func> {
+    /// Creates with filtering `func`, following `sink`.
+    pub fn new(func: Func, sink: Si) -> Self {
+        Self { sink, func }
+    }
+}
+
+impl<Si, Func, Item> Sink<Item> for Filter<Si, Func>
+where
+    Si: Sink<Item>,
+    Func: FnMut(&Item) -> bool,
+{
+    type Error = Si::Error;
+
+    fn poll_ready(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Result<(), Self::Error>> {
+        self.project().sink.poll_ready(cx)
+    }
+    fn start_send(self: Pin<&mut Self>, item: Item) -> Result<(), Self::Error> {
+        let this = self.project();
+        if (this.func)(&item) {
+            this.sink.start_send(item)
+        } else {
+            Ok(())
+        }
+    }
+    fn poll_flush(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Result<(), Self::Error>> {
+        self.project().sink.poll_flush(cx)
+    }
+    fn poll_close(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Result<(), Self::Error>> {
+        self.project().sink.poll_close(cx)
+    }
+}
