+++ dfir_rs/src/compiled/push/filter_map.rs
@@ -1,0 +1,93 @@
+use std::pin::Pin;
+use std::task::{Context, Poll};
+
+use futures::sink::Sink;
+use pin_project_lite::pin_project;
+
+pin_project! {
+    /// Same as [`std::iterator::FilterMap`] but as a [`Sink`].
+    ///
+    /// Synchronously filter-maps items and sends the outputs to the following sink.
+    #[must_use = "sinks do nothing unless polled"]
+    pub struct FilterMap<Si, Func> {
+        #[pin]
+        sink: Si,
+        func: Func,
+    }
+}
+
+impl<Si, Func> FilterMap<Si, Func> {
+    /// Creates with mapping `func`, following `sink`.
+    pub fn new(func: Func, sink: Si) -> Self {
+        Self { sink, func }
+    }
+}
+
+impl<Si, Func, Item, Out> Sink<Item> for FilterMap<Si, Func>
+where
+    Si: Sink<Out>,
+    Func: FnMut(Item) -> Option<Out>,
+{
+    type Error = Si::Error;
+
+    fn poll_ready(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Result<(), Self::Error>> {
+        self.project().sink.poll_ready(cx)
+    }
+    fn start_send(self: Pin<&mut Self>, item: Item) -> Result<(), Self::Error> {
+        let this = self.project();
+        if let Some(item) = (this.func)(item) {
+            this.sink.start_send(item)
+        } else {
+            Ok(())
+        }
+    }
+    fn poll_flush(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Result<(), Self::Error>> {
+        self.project().sink.poll_flush(cx)
+    }
+    fn poll_close(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Result<(), Self::Error>> {
+        self.project().sink.poll_close(cx)
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use futures::sink::SinkExt;
+    use futures::stream::StreamExt;
+    use tokio::sync::mpsc::channel;
+    use tokio_stream::wrappers::ReceiverStream;
+    use tokio_util::sync::PollSender;
+
+    use super::*;
+
+    #[tokio::test]
+    async fn test_filter_map() {
+        let (out_send, out_recv) = channel(1);
+        let out_send = PollSender::new(out_send);
+        let mut out_recv = ReceiverStream::new(out_recv);
+
+        let mut sink = FilterMap::new(std::convert::identity, out_send);
+
+        let a = tokio::task::spawn(async move {
+            sink.send(Some(0)).await.unwrap();
+            sink.send(Some(1)).await.unwrap();
+            sink.send(None).await.unwrap();
+            sink.send(Some(2)).await.unwrap();
+            sink.send(None).await.unwrap();
+            sink.send(None).await.unwrap();
+            sink.send(Some(3)).await.unwrap();
+            sink.send(None).await.unwrap();
+            sink.send(None).await.unwrap();
+            sink.send(None).await.unwrap();
+            sink.send(Some(4)).await.unwrap();
+            sink.send(Some(5)).await.unwrap();
+            sink.send(None).await.unwrap();
+        });
+        println!("{}", line!());
+        assert_eq!(
+            &[0, 1, 2, 3, 4, 5],
+            &*out_recv.by_ref().collect::<Vec<_>>().await
+        );
+        println!("{}", line!());
+        a.await.unwrap();
+    }
+}
