+++ dfir_rs/src/compiled/push/map.rs
@@ -1,0 +1,47 @@
+use std::pin::Pin;
+use std::task::{Context, Poll};
+
+use futures::sink::Sink;
+use pin_project_lite::pin_project;
+
+pin_project! {
+    /// Same as [`std::iterator::Map`] but as a [`Sink`].
+    ///
+    /// Synchronously maps items and sends the output to the following sink.
+    #[must_use = "sinks do nothing unless polled"]
+    pub struct Map<Si, Func> {
+        #[pin]
+        sink: Si,
+        func: Func,
+    }
+}
+
+impl<Si, Func> Map<Si, Func> {
+    /// Creates with mapping `func`, following `sink`.
+    pub fn new(func: Func, sink: Si) -> Self {
+        Self { sink, func }
+    }
+}
+
+impl<Si, Func, Item, Out> Sink<Item> for Map<Si, Func>
+where
+    Si: Sink<Out>,
+    Func: FnMut(Item) -> Out,
+{
+    type Error = Si::Error;
+
+    fn poll_ready(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Result<(), Self::Error>> {
+        self.project().sink.poll_ready(cx)
+    }
+    fn start_send(self: Pin<&mut Self>, item: Item) -> Result<(), Self::Error> {
+        let this = self.project();
+        let item = (this.func)(item);
+        this.sink.start_send(item)
+    }
+    fn poll_flush(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Result<(), Self::Error>> {
+        self.project().sink.poll_flush(cx)
+    }
+    fn poll_close(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Result<(), Self::Error>> {
+        self.project().sink.poll_close(cx)
+    }
+}
