+++ dfir_rs/src/compiled/push/mod.rs
@@ -1,0 +1,18 @@
+//! Push-based operator helpers, i.e. [`futures::sink::Sink`] helpers.
+
+mod filter;
+mod filter_map;
+mod flat_map;
+mod flatten;
+mod for_each;
+mod inspect;
+mod map;
+mod persist;
+pub use filter::Filter;
+pub use filter_map::FilterMap;
+pub use flat_map::FlatMap;
+pub use flatten::Flatten;
+pub use for_each::ForEach;
+pub use inspect::Inspect;
+pub use map::Map;
+pub use persist::Persist;
